move graphing in utils to analytical.py
