+++ todo.md
@@ -5,8 +5,6 @@
 ## New Stuff
 
 - move utils graph functions to analytical.py
-- add graphs for project primos?
-  - graph of primos up until the date mgiht work?
 - user setting for autoimport every time?
 - add social apps login
 - add wish simulator
